+++ pkg/server/controllers.go
@@ -24,6 +24,7 @@ import (
 	"os"
 	"time"
 
+	kcpclienthelper "github.com/kcp-dev/apimachinery/pkg/client"
 	kcpdynamic "github.com/kcp-dev/client-go/dynamic"
 	kcpkubernetesclientset "github.com/kcp-dev/client-go/kubernetes"
 	kcpmetadata "github.com/kcp-dev/client-go/metadata"
@@ -65,11 +66,11 @@ import (
 	schedulinglocationstatus "github.com/kcp-dev/kcp/pkg/reconciler/scheduling/location"
 	schedulingplacement "github.com/kcp-dev/kcp/pkg/reconciler/scheduling/placement"
 	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/bootstrap"
-	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/clusterworkspace"
 	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/clusterworkspaceshard"
 	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/clusterworkspacetype"
 	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/initialization"
 	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/thisworkspace"
+	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/workspace"
 	"github.com/kcp-dev/kcp/pkg/reconciler/tenancy/workspacedeletion"
 	workloadsapiexport "github.com/kcp-dev/kcp/pkg/reconciler/workload/apiexport"
 	workloadsapiexportcreate "github.com/kcp-dev/kcp/pkg/reconciler/workload/apiexportcreate"
@@ -79,7 +80,6 @@ import (
 	workloadresource "github.com/kcp-dev/kcp/pkg/reconciler/workload/resource"
 	synctargetcontroller "github.com/kcp-dev/kcp/pkg/reconciler/workload/synctarget"
 	"github.com/kcp-dev/kcp/pkg/reconciler/workload/synctargetexports"
-	initializingworkspacesbuilder "github.com/kcp-dev/kcp/pkg/virtual/initializingworkspaces/builder"
 )
 
 func postStartHookName(controllerName string) string {
@@ -371,13 +371,13 @@ func (s *Server) installWorkloadResourceScheduler(ctx context.Context, config *r
 
 func (s *Server) installWorkspaceScheduler(ctx context.Context, config *rest.Config, logicalClusterAdminConfig *rest.Config) error {
 	// NOTE: keep `config` unaltered so there isn't cross-use between controllers installed here.
-	clusterWorkspaceConfig := rest.CopyConfig(config)
-	clusterWorkspaceConfig = rest.AddUserAgent(clusterWorkspaceConfig, clusterworkspace.ControllerName)
-	kcpClusterClient, err := kcpclientset.NewForConfig(clusterWorkspaceConfig)
+	workspaceConfig := rest.CopyConfig(config)
+	workspaceConfig = rest.AddUserAgent(workspaceConfig, workspace.ControllerName)
+	kcpClusterClient, err := kcpclientset.NewForConfig(workspaceConfig)
 	if err != nil {
 		return err
 	}
-	kubeClusterClient, err := kcpclientset.NewForConfig(clusterWorkspaceConfig)
+	kubeClusterClient, err := kcpkubernetesclientset.NewForConfig(workspaceConfig)
 	if err != nil {
 		return err
 	}
@@ -385,7 +385,8 @@ func (s *Server) installWorkspaceScheduler(ctx context.Context, config *rest.Con
 	logicalClusterAdminConfig = rest.CopyConfig(logicalClusterAdminConfig)
 	logicalClusterAdminConfig = rest.AddUserAgent(logicalClusterAdminConfig, clusterworkspaceshard.ControllerName)
 
-	workspaceController, err := clusterworkspace.NewController(
+	workspaceController, err := workspace.NewController(
+		s.CompletedConfig.ShardExternalURL,
 		kcpClusterClient,
 		kubeClusterClient,
 		logicalClusterAdminConfig,
@@ -397,8 +398,8 @@ func (s *Server) installWorkspaceScheduler(ctx context.Context, config *rest.Con
 		return err
 	}
 
-	if err := s.AddPostStartHook(postStartHookName(clusterworkspace.ControllerName), func(hookContext genericapiserver.PostStartHookContext) error {
-		logger := klog.FromContext(ctx).WithValues("postStartHook", postStartHookName(clusterworkspace.ControllerName))
+	if err := s.AddPostStartHook(postStartHookName(workspace.ControllerName), func(hookContext genericapiserver.PostStartHookContext) error {
+		logger := klog.FromContext(ctx).WithValues("postStartHook", postStartHookName(workspace.ControllerName))
 		if err := s.waitForSync(hookContext.StopCh); err != nil {
 			logger.Error(err, "failed to finish post-start-hook")
 			return nil // don't klog.Fatal. This only happens when context is cancelled.
@@ -509,8 +510,8 @@ func (s *Server) installWorkspaceScheduler(ctx context.Context, config *rest.Con
 
 func (s *Server) installThisWorkspace(ctx context.Context, config *rest.Config) error {
 	clusterWorkspaceConfig := rest.CopyConfig(config)
-	clusterWorkspaceConfigWithRoundtripper := rest.AddUserAgent(kcpclienthelper.SetMultiClusterRoundTripper(clusterWorkspaceConfig), thisworkspace.ControllerName)
-	kcpClusterClient, err := kcpclient.NewForConfig(clusterWorkspaceConfigWithRoundtripper)
+	clusterWorkspaceConfig = rest.AddUserAgent(clusterWorkspaceConfig, thisworkspace.ControllerName)
+	kcpClusterClient, err := kcpclientset.NewForConfig(clusterWorkspaceConfig)
 	if err != nil {
 		return err
 	}
